Refactor file structure and import statements in dashboard files components, update file URL in file card component

+++ convex/files.ts
@@ -104,6 +104,14 @@ export const deleteFile = mutation({
       throw new ConvexError("You dont have access to delete this file.");
     }
 
+    const isAdmin =
+      access.user.orgIds.find((org) => org.orgId === access.file.orgId)
+        ?.role === "admin";
+
+    if (!isAdmin) {
+      throw new ConvexError("You must be an admin to delete a file.");
+    }
+
     await ctx.db.delete(args.fileId);
   },
 });
@@ -178,7 +186,8 @@ async function hasAccessToOrg(ctx: QueryCtx | MutationCtx, orgId: string) {
   }
 
   const hasAccess =
-    user.orgIds.includes(orgId) || user.tokenIdentifier.includes(orgId);
+    user.orgIds.some((item) => item.orgId === orgId) ||
+    user.tokenIdentifier.includes(orgId);
 
   if (!hasAccess) {
     return null;
